Files under PhotoPreview namespace.
